pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trayrunner"
version = "1.1.0"
description = "A Python 3 tray application for Linux with customizable command menus"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TrayRunner Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.8"
dependencies = [
    "PyGObject>=3.40.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.7",
    "ruamel.yaml>=0.18",
    "pydantic>=2.6",
    "watchdog>=4.0; platform_system!='Windows'",
]

[project.scripts]
trayrunner = "trayrunner.app:main"
trayrunner-gui = "trayrunner_gui.app:main"

[tool.setuptools.packages.find]
where = ["src", "gui"]
include = ["trayrunner*", "trayrunner_gui*"]

[tool.setuptools.package-data]
"trayrunner" = ["icons/*"]
"trayrunner_gui" = ["assets/*"]
